+++ test/bigtest/tests/create-records-plugin-test.js
@@ -29,6 +29,7 @@ describe('CreateInventoryRecords', () => {
 
     describe('save instance', () => {
       beforeEach(async function () {
+        // instance record
         await plugin.modal.form.issnField('123');
         await plugin.modal.form.isbnField('456');
         await plugin.modal.form.fillTitleField('title');
@@ -37,6 +38,13 @@ describe('CreateInventoryRecords', () => {
         await plugin.modal.form.contributors.fillNameField('name');
         await plugin.modal.form.contributors.selectNameTypeField('Personal name');
         await plugin.modal.form.contributors.makeFirstContributorPrimary();
+        // holdings record
+        await plugin.modal.form.openLocationLookup();
+        await plugin.locationLookup.whenLoaded();
+        await plugin.locationLookup.clickOnLocationBtn();
+        await plugin.locationLookup.chooseFirstLocation();
+        await plugin.locationLookup.clickSaveBtn();
+        await plugin.locationLookup.whenClosed();
         await plugin.modal.clickSaveButton();
       });